+++ CHANGELOG.md
@@ -5,11 +5,19 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+- upcloud_managed_database_redis: `redis_version` property
+
 ### Changed
 
 - **Breaking**, upcloud_managed_database_mysql: changing property `admin_password` or `admin_username` forces resource re-creation
 - **Breaking**, upcloud_managed_database_postgresql: changing property `admin_password` or `admin_username` forces resource re-creation
 - **Breaking**, upcloud_managed_database resources: `title` field is required
+
+### Removed
+
+- **Breaking**, upcloud_managed_database_postgresql: `pg_read_replica` property
+- **Breaking**, upcloud_managed_database_postgresql: `pg_service_to_fork_from` property
 
 ## [3.4.0] - 2024-01-25